requirements.txt
fastapi
pyyaml
jinja2
cryptography
uvicorn[standard]
psutil
